git-svn-id: https://php-rpc.svn.sourceforge.net/svnroot/php-rpc/phprpc_3.0@724 88bff967-a00f-0410-b943-93bbd0141c88

+++ cpp/common.hpp
@@ -411,7 +411,7 @@ namespace phprpc
 
 	struct any_hash
 	{
-		uint operator()(const any & value) const
+		size_t operator()(const any & value) const
 		{
 			const std::type_info & vtype = value.type();
